# schema/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Wealth API
servers:
  - url: /api
paths:
  /users:
    post:
      summary: Create a user
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
  /users/contact:
    post:
      summary: Create a user
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - name
                - message
              properties:
                name:
                  type: string
                email:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
  /users/check:
    post:
      summary: Verify an email exists in database
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
  /users/check-verification-token:
    post:
      summary: Check verification token matches and verify account
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - token
                - userID
              properties:
                token:
                  type: string
                userID:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
...
  /portfolios:
    get:
      summary: Fetch all users portfolios
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Create a new portfolio
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - public
                - name
              properties:
                name:
                  type: string
                public:
                  type: boolean
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
...
  /portfolios/{portfolioID}/settings:
    get:
      summary: Fetch a portfolio's settings
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
    put:
      summary: Edit portfolios settings
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - private
                - defaultAssetType
                - reminderEmailPeriod
                - summaryEmailPeriod
              properties:
                private:
                  type: boolean
                reminderEmailPeriod:
                  type: string
                summaryEmailPeriod:
                  type: string
                defaultAssetType:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
...
  /alerts:
    get:
      summary: Get users alerts
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Add alerts
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              required:
                - assetType
                - condition
                - price
                - symbol
                - destination
                - destinationValue
              properties:
                assetType:
                  type: string
                condition:
                  type: string
                price:
                  type: number
                symbol:
                  type: string
                destination:
                  type: string
                destinationValue:
                  type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete alert
      parameters:
        - name: alertID
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: object